Removing the feature of using the actual panel's bg color as the background-optimizing color, going back to a simpler solution in which there's a single, global background color that gets optimized.
Reconciling the bitmask in face of multiple, partially-overlapping rectangles that may have their visibility turned on-off was causing bugs and it is probably more complicated to fix than the feature would be worth.

In practice, it means that the (arguably, rarely needed) UIs that use a lot of large, colored areas (in addition to the common background) might benefit less from the background optimization, causing the colorful areas to be redundantly redrawn more often.