New drawing algorithm.
The old algorithm would figure out, at the level of each panel, the siblings that might cover each other, and use a union rect to exclude them.

The new algorithm maintains a global set of excluded rectangles, and draws items from top to bottom, adding them to the exclusion set.

The exclusion set has some basic optimizations to 'fold' the rectangles in the exclusion set. In particular, adding the panel's bounding box removes the bounding box of all children.

The new algorithm enables future implementation of 'floating' children, i.e. those that are not clipped to their parent's bounds.